Remove the whole onionkey module, and generate onion keys using stem 1.7.0
